extract r2 object commands
